BPA: Better inputs
Take ballot votes during the audit test in from the user to simulate the
API call back and forth with user-input.

Changed the `bravo` function to take one parameter that is a recordclass
(mutable namedtuple) with all of the necessary input parameters.
Eventually, we will change this to an object handed over via an API
call.

Signed-off-by: Ethan Madison <ethan@ethanmad.com>